Remove dependency on utils package
There's some odd behavior here. Removing the depenendcy by moving the file
watchers into core.
