+++ packages/salesforcedx-vscode-apex-debugger/src/index.ts
@@ -266,56 +266,13 @@ function registerFileWatchers(): vscode.Disposable {
   trgWatcher.onDidChange(uri => notifyDebuggerSessionFileChanged());
   trgWatcher.onDidCreate(uri => notifyDebuggerSessionFileChanged());
   trgWatcher.onDidDelete(uri => notifyDebuggerSessionFileChanged());
-  const isvAuthWatcher = vscode.workspace.createFileSystemWatcher(
-    '**/.sfdx/sfdx-config.json'
-  );
-  isvAuthWatcher.onDidChange(uri => notifyIsvAuthChanged());
-  isvAuthWatcher.onDidCreate(uri => notifyIsvAuthChanged());
-  isvAuthWatcher.onDidDelete(uri => notifyIsvAuthChanged());
-  return vscode.Disposable.from(clsWatcher, trgWatcher, isvAuthWatcher);
+  return vscode.Disposable.from(clsWatcher, trgWatcher);
 }
 
 function notifyDebuggerSessionFileChanged(): void {
   if (vscode.debug.activeDebugSession) {
     vscode.debug.activeDebugSession.customRequest(HOTSWAP_REQUEST);
   }
-}
-
-async function notifyIsvAuthChanged() {
-  if (
-    vscode.workspace.workspaceFolders instanceof Array &&
-    vscode.workspace.workspaceFolders.length > 0
-  ) {
-    const forceConfig = await new ForceConfigGet().getConfig(
-      vscode.workspace.workspaceFolders[0].uri.fsPath,
-      SFDX_CONFIG_ISV_DEBUGGER_SID,
-      SFDX_CONFIG_ISV_DEBUGGER_URL
-    );
-    const isvDebuggerSid = forceConfig.get(SFDX_CONFIG_ISV_DEBUGGER_SID);
-    const isvDebuggerUrl = forceConfig.get(SFDX_CONFIG_ISV_DEBUGGER_URL);
-    if (
-      typeof isvDebuggerSid !== 'undefined' &&
-      typeof isvDebuggerUrl !== 'undefined'
-    ) {
-      // set auth context
-      GlobalCliEnvironment.environmentVariables.set(
-        ENV_SFDX_DEFAULTUSERNAME,
-        isvDebuggerSid
-      );
-      GlobalCliEnvironment.environmentVariables.set(
-        ENV_SFDX_INSTANCE_URL,
-        isvDebuggerUrl
-      );
-      console.log(
-        'Configured SFDX_DEFAULTUSERNAME and SFDX_INSTANCE_URL for ISV Project Authentication'
-      );
-      return;
-    }
-  }
-
-  // reset any auth
-  GlobalCliEnvironment.environmentVariables.delete(ENV_SFDX_DEFAULTUSERNAME);
-  GlobalCliEnvironment.environmentVariables.delete(ENV_SFDX_INSTANCE_URL);
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -329,7 +286,6 @@ export function activate(context: vscode.ExtensionContext) {
       new ApexDebuggerConfigurationProvider()
     )
   );
-  notifyIsvAuthChanged();
 }
 
 export function deactivate() {
